+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 **Requires PlatformIO Core 5.1 or above**
 
+* Show active project in statusbar (issue [#2276](https://github.com/platformio/platformio-vscode-ide/issues/2276))
 * Allowed passing custom base URL of the Python Package Index using new `customPyPiIndexUrl` setting
 * Open ["platformio.ini" configuration file](https://docs.platformio.org/page/projectconf/index.html) from newly added project (if there are no other active editors) (issue [#2263](https://github.com/platformio/platformio-vscode-ide/issues/2263))
 * Updated PlatformIO Core installer to v1.0.0 ([release notes](https://github.com/platformio/platformio-core-installer/releases/tag/v1.0.0))